# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Books API",
    "version": "1.0.0",
    "description": "A simple API for working with books"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "summary": "Get all books",
        "tags": [
          "Books"
        ],
        "responses": {
          "200": {
            "description": "Books returned successfully"
          },
          "500": {
            "description": "Unable to retrieve books"
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "summary": "Get one book by id",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The custom book id, such as b1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book returned successfully"
          },
          "404": {
            "description": "Book not found"
          },
          "500": {
            "description": "Unable to retrieve book"
          }
        }
      }
    },
    "/authors": {
      "get": {
        "summary": "Get all authors",
        "tags": [
          "Authors"
        ],
        "responses": {
          "200": {
            "description": "Authors returned successfully"
          },
          "500": {
            "description": "Unable to retrieve authors"
          }
        }
      },
      "post": {
        "summary": "Create an author",
        "tags": [
          "Authors"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "birthYear": {
                    "type": "integer"
                  },
                  "id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created an author"
          },
          "400": {
            "description": "Invalid author details or ID already exist"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "summary": "Get author by id",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "the custom author id, such as a1, a2 and so on",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Author returned successfully"
          },
          "400": {
            "description": "Invalid ID"
          },
          "404": {
            "description": "Author not found"
          },
          "500": {
            "description": "Unable to retrieve authors"
          }
        }
      },
      "put": {
        "summary": "Update an author's details",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "the custom author id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "birthYear": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Author updated successfully"
          },
          "400": {
            "description": "required fields missing"
          },
          "500": {
            "description": "An internal error occured while updating author, try again later"
          }
        }
      },
      "delete": {
        "summary": "delete an author",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Delete an author using the custom ID such as, a1 a2",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted an author successfully"
          },
          "404": {
            "description": "Author does not exist"
          },
          "500": {
            "description": "An internal error occured please try again later"
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}